Place images into subfolders
I suspected Windows was slowing down the download process when the number of
files in an image folder go into the 50k plus range, as this was suggested in
several forums.
To prevent thae potential of that slow down, I added this code to distribute
the images into subfolders based on the first two characters of their random
name string.