Reduce the number of concurrent download processes
While doanloading, I repeatedly had my laptop freeze.
I suspect the problem was the memory was filled up, but I don't have any messages
or logs to confirm that.
As it is, reducing the number of processes is very easy, so it is worth trying
as a solution.
So far so good.