Create subsample with a limited number of classes
Due to the high number of classes in the training set (15000), simply doing a random
sample across the whole set resulted in single member classes nearly everytime.
Now, I restrict the sampling to k randomly selected classes, and those are choosen
in the while loop to be certain to have enough images, ie train + valid number of
images.